Add PropTypeTable (#54)
## Overview
This PR improves the documentation within the demo site by displaying information about each component's props within a table.

This is great for documentation because we do not have to manually keep it in sync with the component. When the component's `propTypes` and `defaultProps` update, so will the `<PropTypeTable />`, automatically.

## Implementation details
It's only capable of detecting simple prop types, like `PropTypes.string`, not the kind that are returned by a function, like `PropTypes.arrayOf(PropTypes.string)`. Still, it is very helpful for the majority of cases, even just to see which props are available and required.

In addition, I've added a non-standard `propDescriptions` static property that can be used to write documentation for each prop _in the context of the component itself_. These descriptions are displayed in the table, in addition to whether or not the prop is required.

## Screenshot
<img width="1620" alt="screen shot 2017-08-10 at 5 28 38 pm" src="https://user-images.githubusercontent.com/2024396/29193570-e9901aac-7df3-11e7-85fb-bb1e0c0a389c.png">

## Questions
1. What do you think of `propDescriptions`? Is it a bad idea to have something non-standard attached to a component?
2. Is there any way to detect more prop types, like `arrayOf()`, `oneOf()`, etc.? So far, [the only implementation](https://github.com/storybooks/storybook/tree/master/addons/info) of such a table that I have seen include that information relied on a babel plugin to inject debug information into the components themselves. I would like to avoid that.

---

**Commits:**

* Add PropTypeTable

* add propTypeTable for Avatar component

* add propTypeTable for Carousel and Checkbox components

* add propTypeTable for Header component

* add propTypeTable for Icon component

* add propTypeTable for Input component

* fix object representation in table

* add propTypeTable for RadioGroup component and give noops correct name

* move noop to utils

* add propTypeTable for Select component

* add propTypeTable for Tag component

* fix propType loop

* add propTypeTable for Text component