Deploy Live Demo Site (#49)
## Overview
This PR makes it possible to have a published demo website like http://quartz.vhx.tv for quartz-react.

I've used firebase for hosting, mainly because it was free and not too hard to set up. But we can use any static file hosting provider, or even just reuse Crystal.

The purpose of this is to make it easy to view the documentation for the quartz-react components without having to start the demo server locally. You can now just go to: https://vhx-quartz-react.firebaseapp.com/

I used my personal vimeo email address to create the account for hosting. If we don't end up switching providers, I should be able to just transfer ownership of the demo site instead.

## Other changes
Since the firebase server uses SSL and our assets (I think) do not, I was not able to fetch them cross-origin. For that reason, I've just copied those stylesheets to the `/public/css` folder.

---

**Commits:**

* set up firebase hosting

* fix cors issue

* actually fix cors issue

* add deploy script

* cleanup